refactor(OmokGameState): 게임 승리를 판단하는 곳에서 인터페이스 형식으로 승리 판단하는 객체를 받기
