+++ src/main/kotlin/domain/OmokGameState.kt
@@ -1,19 +1,23 @@
 package domain
 
 import domain.board.Board
+import domain.library.result.FiveStoneWinningReferee
 import domain.library.result.WinningReferee
 import domain.stone.Color
 
-sealed class OmokGameState(open val winningColor: Color?) {
+sealed class OmokGameState {
 
-    object Running : OmokGameState(null)
+    object Running : OmokGameState()
 
-    data class End(override val winningColor: Color) : OmokGameState(winningColor)
+    data class End(val winningColor: Color) : OmokGameState()
 
     companion object {
-        private val winningReferee: WinningReferee = WinningReferee()
 
-        fun valueOf(currentBoard: Board, color: Color): OmokGameState {
+        fun valueOf(
+            currentBoard: Board,
+            color: Color,
+            winningReferee: WinningReferee = FiveStoneWinningReferee()
+        ): OmokGameState {
             if (winningReferee.checkWin(currentBoard.placedStones, color)) {
                 return End(color)
             }
